- many small anti-warning improvements. - MinGW Win32 ggnfs-lasive4I12e.exe now doesn't print these "Not enough polynominal in mpqs XXX" warnings. WOW!!!
git-svn-id: https://ggnfs.svn.sourceforge.net/svnroot/ggnfs/trunk@139 80ef69ca-9341-0410-9d60-8acc7774fa31

+++ src/lasieve4/mpqs.c
@@ -1980,7 +1980,7 @@ zeitb(10);
 }
 
 
-size_t mpqs_factor(mpz_t N, long max_bits, mpz_t **factors)
+size_t mpqs_factor(mpz_t N, size_t max_bits, mpz_t **factors)
 {
   size_t err;
